refactor : LottoCount, LottoCountCalculator model로 분리
