+++ docs/lottogame/step_two_feature_list.md
@@ -20,7 +20,7 @@
 - [ ] : 수동 로또 생성기 추가 구현
 - [ ] : 수동 로또 만드는 부분 (구매 금액 >= 수동 로또 수 * 로또 가격) 누구에게 책임을 묻지? - 수동 로또 생성기가 하는게 맞음
 - [x] : 자동 로또 생성기 네이밍 변경 (LottieGenerator -> AutoLottieGenerator)
-- [ ] : input : 로또 구매 금액 -> output : 구매할 수 있는 로또 count 를 반환하는 객체를 만들자
+- [x] : LottoCountCalculator 구현 - 돈(cost)으로 구매할 수 있는 로또 count 를 반환하는 객체
 
 # PR에 적을 질문 내용
 기존 코드를 고치기보다는 새로운 코드를 작성하여 리팩토링하는 방식으로 진행했습니다!
